// src/app/shared/store/room/rooms.action.types.ts
export enum RoomsActionTypes {
    OnRoom = "[Room Action] On Room",
    OnRoomSuccess = "[Room Action] On Room Success",
    OnRoomFailure = "[Room Action] On Room Failure",

    OnJoinRoom = "[JoinRoom Action] On JoinRoom",
    OnJoinRoomSuccess = "[JoinRoom Action] On JoinRoom Success",
    OnJoinRoomFailure = "[JoinRoom Action] On JoinRoom Failure",

    OnDeleteRoom = "[DeleteRoom Action] On DeleteRoom",
    OnDeleteRoomSuccess = "[DeleteRoom Action] On DeleteRoom Success",
    OnDeleteRoomFailure = "[DeleteRoom Action] On DeleteRoom Failure",
}